+++ lib/addons.bash
@@ -794,6 +794,8 @@ function addons_install_update {
                                  that was just fetched from repository,
                                  and is not yet present in Odoo database
         --skip-errors          - Do not fail on single DB. Useful in case of update of multiple databases.
+        --show-log-on-error    - Show last 100 lines of log on error.
+        --show-log-on-error=42 - Show last 42 lines of log on error.
     ";
     local need_start;
     local update_addons_list=0;
@@ -860,6 +862,12 @@ function addons_install_update {
             --skip-errors)
                 local skip_errors=1;
             ;;
+            --show-log-on-error)
+                local show_log_on_error=100;
+            ;;
+            --show-log-on-error=*)
+                local show_log_on_error=${key#*=};
+            ;;
             -h|--help|help)
                 echo "$usage";
                 return 0;
@@ -917,16 +925,23 @@ function addons_install_update {
     for db in "${dbs[@]}"; do
         if addons_install_update_internal "$cmd" "$db" "${todo_addons[@]}"; then
             echoe -e "${LBLUEC}${cmd} for ${YELLOWC}$db${LBLUEC}:${NC} ${GREENC}OK${NC}";
-        elif [ -n "$skip_errors" ]; then
-            errored_dbs+=( "$db" );
-            res=1;
         else
-            echoe -e "${LBLUEC}${cmd} for ${YELLOWC}$db${LBLUEC}:${NC} ${REDC}FAIL${NC}";
-            if [ -n "$open_logs" ]; then
-                server_log;
+            if [ -n "$show_log_on_error" ]; then
+                echoe -e "${REDC}ERROR${NC}: update of database ${YELLOWC}${db}${NC} got error. See last lines in log below:";
+                >&2 tail --lines="$show_log_on_error" "$LOG_FILE";
             fi
-            res=1;
-            break
+
+            if [ -n "$skip_errors" ]; then
+                errored_dbs+=( "$db" );
+                res=1;
+            else
+                echoe -e "${LBLUEC}${cmd} for ${YELLOWC}$db${LBLUEC}:${NC} ${REDC}FAIL${NC}";
+                if [ -n "$open_logs" ]; then
+                    server_log;
+                fi
+                res=1;
+                break
+            fi
         fi
     done